requirements.txt
asgiref==3.8.1
Django==4.2.16
django-jazzmin==3.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
packaging==24.2
Pillow==10.0.0
PyJWT==2.9.0
python-dotenv==1.0.0
shortuuid==1.0.13
sqlparse==0.5.2
typing-extensions==4.12.2
httpx